[DEV-1550] Refactor "alias" handling for PayOut
