+++ src/subdomains/supporting/payout/strategies/payout/__tests__/payout-jellyfish.strategy.spec.ts
@@ -1,17 +1,18 @@
 import { mock } from 'jest-mock-extended';
-import { NotificationService } from 'src/subdomains/supporting/notification/services/notification.service';
-import { Asset } from 'src/shared/models/asset/asset.entity';
+import { Blockchain } from 'src/integration/blockchain/shared/enums/blockchain.enum';
 import { createCustomAsset } from 'src/shared/models/asset/__mocks__/asset.entity.mock';
-import { PayoutOrder, PayoutOrderStatus } from '../../../entities/payout-order.entity';
+import { Asset, AssetType } from 'src/shared/models/asset/asset.entity';
+import { DfxLogger } from 'src/shared/services/dfx-logger';
+import { NotificationService } from 'src/subdomains/supporting/notification/services/notification.service';
 import {
   createCustomPayoutOrder,
   createDefaultPayoutOrder,
 } from '../../../entities/__mocks__/payout-order.entity.mock';
+import { PayoutOrder, PayoutOrderStatus } from '../../../entities/payout-order.entity';
 import { FeeResult } from '../../../interfaces';
 import { PayoutOrderRepository } from '../../../repositories/payout-order.repository';
 import { PayoutDeFiChainService } from '../../../services/payout-defichain.service';
 import { JellyfishStrategy } from '../impl/base/jellyfish.strategy';
-import { DfxLogger } from 'src/shared/services/dfx-logger';
 
 describe('PayoutJellyfishStrategy', () => {
   let strategy: PayoutJellyfishStrategyWrapper;
@@ -242,6 +243,9 @@ describe('PayoutJellyfishStrategy', () => {
 
 class PayoutJellyfishStrategyWrapper extends JellyfishStrategy {
   protected readonly logger = new DfxLogger(PayoutJellyfishStrategyWrapper);
+
+  blockchain = Blockchain.DEFICHAIN;
+  assetType = AssetType.COIN;
 
   constructor(
     notificationService: NotificationService,
