+++ src/subdomains/supporting/payout/strategies/payout/impl/defichain-coin.strategy.ts
@@ -1,18 +1,22 @@
 import { Injectable } from '@nestjs/common';
-import { PayoutOrderContext, PayoutOrder } from '../../../entities/payout-order.entity';
+import { Blockchain } from 'src/integration/blockchain/shared/enums/blockchain.enum';
+import { Asset, AssetType } from 'src/shared/models/asset/asset.entity';
+import { AssetService } from 'src/shared/models/asset/asset.service';
+import { DfxLogger } from 'src/shared/services/dfx-logger';
+import { NotificationService } from 'src/subdomains/supporting/notification/services/notification.service';
+import { PayoutOrder, PayoutOrderContext } from '../../../entities/payout-order.entity';
+import { FeeResult } from '../../../interfaces';
 import { PayoutOrderRepository } from '../../../repositories/payout-order.repository';
 import { PayoutGroup } from '../../../services/base/payout-jellyfish.service';
 import { PayoutDeFiChainService } from '../../../services/payout-defichain.service';
 import { JellyfishStrategy } from './base/jellyfish.strategy';
-import { NotificationService } from 'src/subdomains/supporting/notification/services/notification.service';
-import { AssetService } from 'src/shared/models/asset/asset.service';
-import { FeeResult } from '../../../interfaces';
-import { Asset } from 'src/shared/models/asset/asset.entity';
-import { DfxLogger } from 'src/shared/services/dfx-logger';
 
 @Injectable()
 export class DeFiChainCoinStrategy extends JellyfishStrategy {
   protected readonly logger = new DfxLogger(DeFiChainCoinStrategy);
+
+  blockchain = Blockchain.DEFICHAIN;
+  assetType = AssetType.COIN;
 
   constructor(
     notificationService: NotificationService,
