+++ src/subdomains/supporting/payout/strategies/prepare/__tests__/prepare.registry.spec.ts
@@ -1,0 +1,120 @@
+import { mock } from 'jest-mock-extended';
+import { Blockchain } from 'src/integration/blockchain/shared/enums/blockchain.enum';
+import { createCustomAsset } from 'src/shared/models/asset/__mocks__/asset.entity.mock';
+import { AssetService } from 'src/shared/models/asset/asset.service';
+import { DexService } from 'src/subdomains/supporting/dex/services/dex.service';
+import { PayoutOrderRepository } from '../../../repositories/payout-order.repository';
+import { PayoutDeFiChainService } from '../../../services/payout-defichain.service';
+import { ArbitrumStrategy } from '../impl/arbitrum.strategy';
+import { PrepareStrategyRegistry } from '../impl/base/prepare.strategy-registry';
+import { BitcoinStrategy } from '../impl/bitcoin.strategy';
+import { BscStrategy } from '../impl/bsc.strategy';
+import { DeFiChainStrategy } from '../impl/defichain.strategy';
+import { EthereumStrategy } from '../impl/ethereum.strategy';
+import { OptimismStrategy } from '../impl/optimism.strategy';
+
+describe('PrepareStrategyRegistry', () => {
+  let bitcoinStrategy: BitcoinStrategy;
+  let defichainStrategy: DeFiChainStrategy;
+  let ethereumStrategy: EthereumStrategy;
+  let bscStrategy: BscStrategy;
+  let arbitrumStrategy: ArbitrumStrategy;
+  let optimismStrategy: OptimismStrategy;
+
+  let registry: PrepareStrategyRegistryWrapper;
+
+  beforeEach(() => {
+    bitcoinStrategy = new BitcoinStrategy(mock<AssetService>(), mock<PayoutOrderRepository>());
+    defichainStrategy = new DeFiChainStrategy(
+      mock<AssetService>(),
+      mock<DexService>(),
+      mock<PayoutDeFiChainService>(),
+      mock<PayoutOrderRepository>(),
+    );
+    ethereumStrategy = new EthereumStrategy(mock<AssetService>(), mock<PayoutOrderRepository>());
+    bscStrategy = new BscStrategy(mock<AssetService>(), mock<PayoutOrderRepository>());
+    arbitrumStrategy = new ArbitrumStrategy(mock<AssetService>(), mock<PayoutOrderRepository>());
+    optimismStrategy = new OptimismStrategy(mock<AssetService>(), mock<PayoutOrderRepository>());
+
+    registry = new PrepareStrategyRegistryWrapper(
+      bitcoinStrategy,
+      defichainStrategy,
+      ethereumStrategy,
+      bscStrategy,
+      arbitrumStrategy,
+      optimismStrategy,
+    );
+  });
+
+  describe('#getPrepareStrategy(...)', () => {
+    describe('getting strategy by Asset', () => {
+      it('gets BITCOIN strategy for BITCOIN', () => {
+        const strategy = registry.getPrepareStrategy(createCustomAsset({ blockchain: Blockchain.BITCOIN }));
+
+        expect(strategy).toBeInstanceOf(BitcoinStrategy);
+      });
+
+      it('gets ETHEREUM strategy', () => {
+        const strategy = registry.getPrepareStrategy(createCustomAsset({ blockchain: Blockchain.ETHEREUM }));
+
+        expect(strategy).toBeInstanceOf(EthereumStrategy);
+      });
+
+      it('gets BSC strategy', () => {
+        const strategy = registry.getPrepareStrategy(createCustomAsset({ blockchain: Blockchain.BINANCE_SMART_CHAIN }));
+
+        expect(strategy).toBeInstanceOf(BscStrategy);
+      });
+
+      it('gets DEFICHAIN strategy for DEFICHAIN', () => {
+        const strategy = registry.getPrepareStrategy(createCustomAsset({ blockchain: Blockchain.DEFICHAIN }));
+
+        expect(strategy).toBeInstanceOf(DeFiChainStrategy);
+      });
+
+      it('gets ARBITRUM strategy for ARBITRUM', () => {
+        const strategy = registry.getPrepareStrategy(createCustomAsset({ blockchain: Blockchain.ARBITRUM }));
+
+        expect(strategy).toBeInstanceOf(ArbitrumStrategy);
+      });
+
+      it('gets OPTIMISM strategy for OPTIMISM', () => {
+        const strategy = registry.getPrepareStrategy(createCustomAsset({ blockchain: Blockchain.OPTIMISM }));
+
+        expect(strategy).toBeInstanceOf(OptimismStrategy);
+      });
+
+      it('fails to get strategy for non-supported Blockchain', () => {
+        const testCall = () =>
+          registry.getPrepareStrategy(createCustomAsset({ blockchain: 'NewBlockchain' as Blockchain }));
+
+        expect(testCall).toThrow();
+        expect(testCall).toThrowError('No PrepareStrategy found. Blockchain: NewBlockchain');
+      });
+    });
+  });
+});
+
+class PrepareStrategyRegistryWrapper extends PrepareStrategyRegistry {
+  constructor(
+    bitcoinStrategy: BitcoinStrategy,
+    defichainStrategy: DeFiChainStrategy,
+    ethereumStrategy: EthereumStrategy,
+    bscStrategy: BscStrategy,
+    arbitrumStrategy: ArbitrumStrategy,
+    optimismStrategy: OptimismStrategy,
+  ) {
+    super();
+
+    this.addStrategy(bitcoinStrategy, Blockchain.BITCOIN);
+    this.addStrategy(defichainStrategy, Blockchain.DEFICHAIN);
+    this.addStrategy(ethereumStrategy, Blockchain.ETHEREUM);
+    this.addStrategy(bscStrategy, Blockchain.BINANCE_SMART_CHAIN);
+    this.addStrategy(arbitrumStrategy, Blockchain.ARBITRUM);
+    this.addStrategy(optimismStrategy, Blockchain.OPTIMISM);
+  }
+
+  getStrategy(blockchain: Blockchain) {
+    return super.getStrategy(blockchain);
+  }
+}
