+++ src/subdomains/supporting/payout/strategies/prepare/impl/bitcoin.strategy.ts
@@ -1,4 +1,5 @@
 import { Injectable } from '@nestjs/common';
+import { Blockchain } from 'src/integration/blockchain/shared/enums/blockchain.enum';
 import { Asset } from 'src/shared/models/asset/asset.entity';
 import { AssetService } from 'src/shared/models/asset/asset.service';
 import { PayoutOrderRepository } from '../../../repositories/payout-order.repository';
@@ -6,7 +7,9 @@ import { AutoConfirmStrategy } from './base/auto-confirm.strategy';
 
 @Injectable()
 export class BitcoinStrategy extends AutoConfirmStrategy {
-  constructor(protected readonly assetService: AssetService, payoutOrderRepo: PayoutOrderRepository) {
+  blockchain = Blockchain.BITCOIN;
+
+  constructor(private readonly assetService: AssetService, payoutOrderRepo: PayoutOrderRepository) {
     super(payoutOrderRepo);
   }
