+++ src/subdomains/supporting/payout/strategies/prepare/impl/optimism.strategy.ts
@@ -1,4 +1,5 @@
 import { Injectable } from '@nestjs/common';
+import { Blockchain } from 'src/integration/blockchain/shared/enums/blockchain.enum';
 import { Asset } from 'src/shared/models/asset/asset.entity';
 import { AssetService } from 'src/shared/models/asset/asset.service';
 import { PayoutOrderRepository } from '../../../repositories/payout-order.repository';
@@ -6,7 +7,9 @@ import { EvmStrategy } from './base/evm.strategy';
 
 @Injectable()
 export class OptimismStrategy extends EvmStrategy {
-  constructor(protected readonly assetService: AssetService, payoutOrderRepo: PayoutOrderRepository) {
+  blockchain = Blockchain.OPTIMISM;
+
+  constructor(private readonly assetService: AssetService, payoutOrderRepo: PayoutOrderRepository) {
     super(payoutOrderRepo);
   }
